fix api injection in Cypress tests
